refactor: 将用户余额相关类型从double改为BigDecimal
更新了DataInitializer里的ensureFields方法参数，以及UserDetailsSuccess DTO的余额字段和构造函数参数类型，避免浮点精度问题

+++ server/src/main/java/me/code/springboot_postgres/config/DataInitializer.java
@@ -84,7 +84,7 @@ public class DataInitializer implements CommandLineRunner {
         }
     }
 
-    private void ensureFields(User user, User.LegacyRole legacyRole, double balance, boolean isProtected) {
+    private void ensureFields(User user, User.LegacyRole legacyRole, BigDecimal balance, boolean isProtected) {
         boolean needsUpdate = false;
 
         if (user.getLegacyRole() != legacyRole) {
